+++ packages/koin_test/lib/src/error/error.dart
@@ -1,5 +1,5 @@
 /*
- * Copyright 2017-2018 the original author or authors.
+ * Copyright 2017-2019 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -22,6 +22,10 @@
  * Ported to Dart from Kotlin by:
  * @author - Pedro Bissonho 
  */
+
+///
+/// Exception thrown when a Koin Definition not run.
+///
 abstract class BrokenDefinitionException implements Exception {
   final String msg;
   BrokenDefinitionException(this.msg);