+++ packages/koin_test/CHANGELOG.md
@@ -1,3 +1,15 @@
+## 0.5.0
+
+- Added functions to improve koin tests.
+    - `koinTest` Configures the testing environment to automatically start and `close Koin.
+    - `koinSetUp` Register a setUp function that starts koin before tests.
+    - `koinTearDown` Register a tearDown function that close Koin afters tests.
+    - `inject` Lazy inject an instance from Koin in the test environment.
+    - `get` Get an instance from Koin in the test environment.
+
+
+- Support for koin 0.2.0
+
 ## 0.2.0
 
 - Support for koin 0.2.0